Feature/fehler

+++ basic_calculations.py
@@ -1,5 +1,5 @@
-def addieren(a: int , b: int) -> int:
-    return a + b
+def addieren(a: float , b: float) -> int:
+    return float(a) + float(b)
 
 def subtrahieren(a: int, b: int) -> int:
     return a - b
